Move get_waypoint_coords to graph
